vec3, vec4 and mat4 converted
